alterações para float ao invés de int

+++ Verificacao.py
@@ -10,7 +10,7 @@ class Verificacao:
       #ele ignora os dados que não podem ser convertidos em números.
       for item in entrada:
         try: 
-         numero = int(item)
+         numero = float(item)
          numeros.append(numero) 
         except:
          pass
